requirements.txt
alembic==1.16.4 ; python_version >= "3.13" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.13" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.13" and python_version < "4.0"
argon2-cffi-bindings==21.2.0 ; python_version >= "3.13" and python_version < "4.0"
argon2-cffi==23.1.0 ; python_version >= "3.13" and python_version < "4.0"
certifi==2025.7.9 ; python_version >= "3.13" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.13" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.13" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.13" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
dnspython==2.7.0 ; python_version >= "3.13" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.13" and python_version < "4.0"
fastapi-cli==0.0.8 ; python_version >= "3.13" and python_version < "4.0"
fastapi-cloud-cli==0.1.2 ; python_version >= "3.13" and python_version < "4.0"
fastapi==0.116.0 ; python_version >= "3.13" and python_version < "4.0"
greenlet==3.2.3 ; python_version == "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.16.0 ; python_version >= "3.13" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.13" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.13" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.13" and python_version < "4.0"
idna==3.10 ; python_version >= "3.13" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.13" and python_version < "4.0"
mako==1.3.10 ; python_version >= "3.13" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.13" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.13" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.13" and python_version < "4.0"
psycopg-binary==3.2.9 ; python_version >= "3.13" and python_version < "4.0" and implementation_name != "pypy"
psycopg==3.2.9 ; python_version >= "3.13" and python_version < "4.0"
pwdlib==0.2.1 ; python_version >= "3.13" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.13" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.13" and python_version < "4.0"
pydantic-settings==2.10.1 ; python_version >= "3.13" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.13" and python_version < "4.0"
pygments==2.19.2 ; python_version >= "3.13" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.13" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.13" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.13" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.13" and python_version < "4.0"
rich-toolkit==0.14.8 ; python_version >= "3.13" and python_version < "4.0"
rich==14.0.0 ; python_version >= "3.13" and python_version < "4.0"
rignore==0.5.1 ; python_version >= "3.13" and python_version < "4.0"
sentry-sdk==2.32.0 ; python_version >= "3.13" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.13" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.13" and python_version < "4.0"
sqlalchemy==2.0.41 ; python_version >= "3.13" and python_version < "4.0"
starlette==0.46.2 ; python_version >= "3.13" and python_version < "4.0"
typer==0.16.0 ; python_version >= "3.13" and python_version < "4.0"
typing-extensions==4.14.1 ; python_version >= "3.13" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.13" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.13" and python_version < "4.0" and sys_platform == "win32"
urllib3==2.5.0 ; python_version >= "3.13" and python_version < "4.0"
uvicorn==0.35.0 ; python_version >= "3.13" and python_version < "4.0"
uvloop==0.21.0 ; python_version >= "3.13" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
watchfiles==1.1.0 ; python_version >= "3.13" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.13" and python_version < "4.0"
